mover to visual studio; node; container node; executablecontainerNode; working

+++ src/fluidControl/executable/containers/BidirectionalSwitch.h
@@ -70,7 +70,7 @@ inline void BidirectionalSwitch::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( BidirectionalSwitch, 1 );
+CEREAL_CLASS_VERSION( BidirectionalSwitch, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
