+++ src/fluidControl/executable/containers/ConvergentSwitch.h
@@ -61,7 +61,7 @@ inline void ConvergentSwitch::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( ConvergentSwitch, 1 );
+CEREAL_CLASS_VERSION( ConvergentSwitch, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
