+++ src/fluidControl/executable/containers/DivergentSwitch.h
@@ -59,7 +59,7 @@ inline void DivergentSwitch::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( DivergentSwitch, 1 );
+CEREAL_CLASS_VERSION( DivergentSwitch, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
