+++ src/fluidControl/executable/containers/FlowContainer.h
@@ -61,7 +61,7 @@ inline void FlowContainer::serialize(Archive& ar, const std::uint32_t version) {
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( FlowContainer, 1 );
+CEREAL_CLASS_VERSION( FlowContainer, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
