+++ src/fluidControl/executable/containers/InletContainer.h
@@ -57,7 +57,7 @@ inline void InletContainer::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( InletContainer, 1 );
+CEREAL_CLASS_VERSION( InletContainer, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
