+++ src/fluidControl/executable/containers/actuators/extras/ODSensor.h
@@ -46,7 +46,7 @@ inline void ODSensor::serialize(Archive& ar, const std::uint32_t version) {
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( ODSensor, 1 );
+CEREAL_CLASS_VERSION( ODSensor, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
