+++ src/testactuator/EvoprogDiscretePump.h
@@ -45,7 +45,7 @@ inline void EvoprogDiscretePump::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( EvoprogDiscretePump, 1 );
+CEREAL_CLASS_VERSION( EvoprogDiscretePump, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
