+++ src/testactuator/EvoprogDummyInjector.h
@@ -41,7 +41,7 @@ inline void EvoprogDummyInjector::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( EvoprogDummyInjector, 1 );
+CEREAL_CLASS_VERSION( EvoprogDummyInjector, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
