+++ src/testactuator/EvoprogOdSensor.h
@@ -49,7 +49,7 @@ inline void EvoprogOdSensor::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( EvoprogOdSensor, 1 );
+CEREAL_CLASS_VERSION( EvoprogOdSensor, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
