+++ src/testactuator/EvoprogSixwayValve.h
@@ -55,7 +55,7 @@ inline void EvoprogSixwayValve::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( EvoprogSixwayValve, 1 );
+CEREAL_CLASS_VERSION( EvoprogSixwayValve, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
