+++ src/testactuator/EvoprogTemperature.h
@@ -47,7 +47,7 @@ inline void EvoprogTemperature::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( EvoprogTemperature, 1 );
+CEREAL_CLASS_VERSION( EvoprogTemperature, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
